Move test-related bin functionality out of `trustfall_core`.
