+++ trustfall_core/Cargo.toml
@@ -13,6 +13,10 @@ rustdoc-args = ["--cfg", "docsrs"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[features]
+default = []
+__private = []
+
 [lib]
 name = "trustfall_core"
 path = "src/lib.rs"
