+++ trustfall_core/src/graphql_query/mod.rs
@@ -3,3 +3,9 @@
 pub(crate) mod directives;
 pub mod error;
 pub(crate) mod query;
+
+// Test-only uses. `#[doc(hidden)]` items are not part of public API
+// and are not subject to semantic versioning rules.
+#[cfg(feature = "__private")]
+#[doc(hidden)]
+pub use query::parse_document;
