+++ trustfall_core/src/util.rs
@@ -1,17 +1,8 @@
 use std::collections::{btree_map, hash_map, BTreeMap, HashMap};
 use std::fmt::{Debug, Display};
 use std::hash::Hash;
-use std::sync::Arc;
 
 use serde::{Deserialize, Serialize};
-
-use crate::ir::Output;
-use crate::{
-    frontend::error::FrontendError,
-    graphql_query::{error::ParseError, query::Query},
-    interpreter::trace::Trace,
-    ir::{FieldValue, IRQuery},
-};
 
 #[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
 pub struct DisplayVec<T>(pub Vec<T>);
@@ -150,61 +141,4 @@ impl<K: Ord, V> BTreeMapTryInsertExt<K, V> for BTreeMap<K, V> {
             btree_map::Entry::Occupied(entry) => Err(BTreeMapOccupiedError { entry, value }),
         }
     }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(crate) struct TestGraphQLQuery {
-    pub(crate) schema_name: String,
-
-    pub(crate) query: String,
-
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub(crate) arguments: BTreeMap<String, FieldValue>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(crate) struct TestParsedGraphQLQuery {
-    pub(crate) schema_name: String,
-
-    pub(crate) query: Query,
-
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub(crate) arguments: BTreeMap<String, FieldValue>,
-}
-
-#[allow(dead_code)]
-pub(crate) type TestParsedGraphQLQueryResult = Result<TestParsedGraphQLQuery, ParseError>;
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(crate) struct TestIRQuery {
-    pub(crate) schema_name: String,
-
-    pub(crate) ir_query: IRQuery,
-
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub(crate) arguments: BTreeMap<String, FieldValue>,
-}
-
-#[allow(dead_code)]
-pub(crate) type TestIRQueryResult = Result<TestIRQuery, FrontendError>;
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(bound = "Vertex: Serialize, for<'de2> Vertex: Deserialize<'de2>")]
-pub(crate) struct TestInterpreterOutputTrace<Vertex>
-where
-    Vertex: Clone + Debug + PartialEq + Eq + Serialize,
-    for<'de2> Vertex: Deserialize<'de2>,
-{
-    pub(crate) schema_name: String,
-
-    pub(crate) trace: Trace<Vertex>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(crate) struct TestInterpreterOutputData {
-    pub(crate) schema_name: String,
-
-    pub(crate) outputs: BTreeMap<Arc<str>, Output>,
-
-    pub(crate) results: Vec<BTreeMap<Arc<str>, FieldValue>>,
 }
